+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Can now enable/disable a layout with `Layout.set_enabled`. New examples 53, 53bis.
 - changing text color in `Text_display` is now possible, using
   `Widget.html`, see exemple 47
 - underlining text in `Text_display` is now possible, using